fix: app stuck generating assistant response (#2001)
